Feature add new metrics (#20)
* Refactoring the creation of the files with the results of metrics

* Added to the energy consumption metric

* Addition of blocking components by quality of transmission and changes in related classes

* Adds metrics for using modulatioon formats

* Correction in the modulation utization metric

* Fixed class FixedRoutes to work with a route file in Gson format

* Small correction in the RandGenerator class

* Modifying the modulation utilization metric to work with translucdnt circuits

+++ src/main/java/network/Regenerators.java
@@ -9,11 +9,11 @@ import java.io.Serializable;
  */
 public class Regenerators implements Serializable {
 
-	private double regenUtilization;
-    private double numRegenerators;
+	private int regenUtilization;
+    private int numRegenerators;
     
     public Regenerators(int numberOfRegenerators){
-    	this.regenUtilization = 0.0;
+    	this.regenUtilization = 0;
     	this.numRegenerators = numberOfRegenerators;
     }
     
@@ -43,28 +43,28 @@ public class Regenerators implements Serializable {
 		return (regenUtilization < numRegenerators);
 	}
 	
-	public double quantRegeneratorsFree(){
+	public int quantRegeneratorsFree(){
 		return numRegenerators - regenUtilization;
 	}
 
 	/**
 	 * @return the regenUtilization
 	 */
-	public double getRegenUtilization() {
+	public int getRegenUtilization() {
 		return regenUtilization;
 	}
 
 	/**
 	 * @return the numRegenerators
 	 */
-	public double getNumRegenerators() {
+	public int getNumRegenerators() {
 		return numRegenerators;
 	}
 
 	/**
 	 * @param numRegenerators the numRegenerators to set
 	 */
-	public void setNumRegenerators(double numRegenerators) {
+	public void setNumRegenerators(int numRegenerators) {
 		this.numRegenerators = numRegenerators;
 	}
